Figured out where to put the search state variables
